Moved Net classes to Configuration namespace
